Implement Events as a linked list.
[js perf](http://jsperf.com/backbone-events-linked-list/18)

Roughly equivalent performance. Using a linked list allows the events to
be mutated by trigger, without skipping an event. This is a **breaking
change**.

Events used to use a linked list (and switched to an array in #1284).
But, it performed [very badly](http://jsperf.com/backbone-events-linked-list/8).
I'm fairly certain that was due to
[L106](https://github.com/jashkenas/backbone/blob/6c3d3838e10d5a070f0d1f604fb5c99ed34c8c93/backbone.js#L106)
recreating the linked list's object every `#on`, and
[L138](https://github.com/jashkenas/backbone/blob/6c3d3838e10d5a070f0d1f604fb5c99ed34c8c93/backbone.js#L138)
calling `#on` to re-bind callbacks inside `#off`.

This fixes #3466 and closes #3463.